+++ utilities.h
@@ -16,7 +16,7 @@ typedef struct {
 
 void readNodeLine(char *buffer, node *nodes, int i);
 void printNodesList(node *nodes, int length);
-unsigned nodesearch(node *first, unsigned long id, unsigned length);
+unsigned nodesearch(node *first, unsigned long id, unsigned long length);
 int searchInSuccessors(node *first, unsigned prev_pos, unsigned long next_id);
 void linkNodes(node *nodes, unsigned prev_pos, unsigned next_pos);
 void computeValences(node *nodes, unsigned long length, int max_valence);
